pyproject.toml
[project]
name = "audio-transcript-bot"
version = "0.1.0"
description = "Telegram bot: forward voice messages, get transcript or summary"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[webhooks]>=21.11,<22",
    "groq>=0.15.0",
    "google-genai>=0.8.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0"]

[tool.ruff]
line-length = 100
target-version = "py311"
